reversing some architecture changes
