+++ prinz/src/main/scala/pl/touk/nussknacker/prinz/model/ModelInstance.scala
@@ -5,7 +5,7 @@ import pl.touk.nussknacker.prinz.util.nussknacker.NKConverter
 
 import scala.concurrent.Future
 
-abstract class ModelInstance(model: Model, private val signatureInterpreter: SignatureInterpreter) {
+abstract class ModelInstance(model: Model, protected val signatureInterpreter: SignatureInterpreter) {
 
   type ModelRunResult = Future[Either[ModelRunException, List[AnyRef]]]
 
@@ -17,26 +17,5 @@ abstract class ModelInstance(model: Model, private val signatureInterpreter: Sig
   def getSignature: ModelSignature = signatureOption match {
     case Some(value) => value
     case None => throw SignatureNotFoundException(this)
-  }
-
-  def validateInput(input: VectorMultimap[String, AnyRef]): Boolean = {
-    val signature = getSignature
-    var result = true
-
-    input.takeWhile(_ => result).foreach(el => {
-      val(k, v) = el
-      val require = signature.getInputValueType(SignatureName(k))
-      val given = NKConverter.toTypingResult(v.getClass.toString)
-
-      if(require.isEmpty) {
-        result = false
-      }
-
-      else if(!given.canBeSubclassOf(require.get.typingResult)) {
-        result = false
-      }
-    })
-
-    result
   }
 }
